Added mutator and accessor to Product
This allows setting and retrieving the price as a decimal (dollars.cents) value and storing in database as integer cents.

Also updated/added tests to check accessor/mutator.

+++ app/Domain/PendingSales/Actions/PricePatchAction.php
@@ -21,10 +21,10 @@ class PricePatchAction
 
     /**
      * @param Product $product
-     * @param int $price
+     * @param float $price
      * @return Product
      */
-    public static function execute(Product $product, int $price): Product
+    public static function execute(Product $product, float $price): Product
     {
         $product->price = $price;
         $product->save();
